go to GRIND MODE FAM
charts are fixed now, and everything's assigned correctly in there. Dialogue's fixed and implemented, I spent like an hour and 45 minutes adjusting the house stage because flixel positioning sucks dick, bastardns (bastard no screwdriver) has been implemented and is used in songs 2 and 3 and the readme's in now with credits n shit. All that's remaining is the production and implementation of the campaign menu sprites and dialogue portraits, and then we should be done!